feat: config jaCoCo and sonar
