+++ .github/workflows/ci-build-checker.yml
@@ -24,6 +24,7 @@ jobs:
     - name: Validate Gradle Wrapper
       uses: gradle/wrapper-validation-action@e6e38bacfdf1a337459f332974bb2327a31aaf4b
     - name: Build and Test
-      uses: gradle/gradle-build-action@2.4.2
+      uses: gradle/gradle-build-action@v2
       with:
         arguments: -S -i build
+        gradle-version: 8.1
