+++ .gitignore
@@ -35,3 +35,4 @@ out/
 
 ### VS Code ###
 .vscode/
+docker/data/**
