+++ Dockerfile
@@ -10,13 +10,6 @@ USER ibeer-user:ibeer-group
 ARG JAR_FILE=build/libs/*jar
 COPY ${JAR_FILE} ibeer-service.jar
 
-ENTRYPOINT [
-            "java", \
-            "-Xms128m", \
-            "-Xmx256m", \
-            "-XX:MetaspaceSize=64m", \
-            "-XX:MaxMetaspaceSize=128m", \
-            "-jar","/ibeer-service.jar" \
-]
+ENTRYPOINT ["java", "-Xms128m", "-Xmx256m", "-XX:MetaspaceSize=64m", "-XX:MaxMetaspaceSize=128m", "-jar", "ibeer-service.jar"]
 
 EXPOSE 8081
