+++ build.gradle
@@ -1,6 +1,8 @@
 plugins {
     id "org.springframework.boot" version "3.1.0"
     id "io.spring.dependency-management" version "1.1.0"
+    id "org.springdoc.openapi-gradle-plugin" version "1.6.0"
+    id "org.sonarqube" version "4.2.1.3168"
     id "application"
     id "jacoco"
     id "java"
@@ -18,6 +20,8 @@ ext {
     set("swaggerVersion", "2.9.2")
     set("openApiVersion", "2.1.0")
     set("hamcrestVersion", "1.3")
+
+    set("sonarExclusions", "com.ipaixao.ibeer.infrastructure.config.*")
 }
 
 group = "com.ipaixao.ibeer"
@@ -43,7 +47,7 @@ dependencies {
     implementation "org.springframework.boot:spring-boot-starter-data-jpa"
     implementation "org.springframework.boot:spring-boot-starter-security"
     implementation "org.springframework.boot:spring-boot-starter-web"
-    developmentOnly 'org.springframework.boot:spring-boot-devtools'
+    developmentOnly "org.springframework.boot:spring-boot-devtools"
 
     implementation "org.springdoc:springdoc-openapi-starter-webmvc-ui:${openApiVersion}"
     implementation "org.apache.commons:commons-collections4:${commonsCollections}"
@@ -65,8 +69,18 @@ dependencies {
     testImplementation "org.springframework.boot:spring-boot-starter-test"
     testImplementation "org.springframework.security:spring-security-test"
 
-    runtimeOnly 'io.micrometer:micrometer-registry-prometheus'
+    runtimeOnly "io.micrometer:micrometer-registry-prometheus"
     runtimeOnly "com.h2database:h2"
+}
+
+application {
+    mainClassName = "com.ipaixao.ibeer.IBeerApplication"
+}
+
+openApi {
+    apiDocsUrl.set("http://localhost:8081/api-docs")
+    outputDir.set(file("$projectDir/docs"))
+    outputFileName.set("swagger.json")
 }
 
 bootBuildImage {
@@ -76,18 +90,59 @@ bootBuildImage {
     imageName = "ipaixao/${applicationName}:${project.version}"
 }
 
-application {
-    mainClassName = "com.ipaixao.ibeer.IBeerApplication"
-}
-
 test {
     useJUnitPlatform()
     finalizedBy jacocoTestReport
 }
+
+check {
+    dependsOn jacocoTestCoverageVerification
+}
+
 jacocoTestReport {
     dependsOn test
     reports {
-        xml.enabled false
-        csv.enabled false
+        xml.required = true
+        csv.required = false
+        html.outputLocation = layout.buildDirectory.dir("jacoco-html-reports")
     }
 }
+jacocoTestCoverageVerification {
+    violationRules {
+        rule {
+            element = "PACKAGE"
+            includes = ["com.ipaixao.ibeer.*"]
+            excludes = [sonarExclusions]
+
+            limit {
+                counter = "INSTRUCTION"
+                value = "COVEREDRATIO"
+                minimum = 0.43
+            }
+
+            limit {
+                counter = "CLASS"
+                value = "MISSEDCOUNT"
+                maximum = 0
+            }
+        }
+    }
+}
+
+sonar {
+    properties {
+        property("sonar.projectKey", "ibeer-service")
+        property("sonar.projectName", "iBeer Service")
+        property("sonar.organization", "com.ipaixao")
+        property("sonar.host.url", "http://localhost:9000")
+        property("sonar.token", "sqp_287ead3f4ad25ba06fee0d4a26d36e2f5d9fb54f")
+
+        property("sonar.coverage.jacoco.xmlReportPaths", "$buildDir/reports/jacoco/test/*.xml")
+        property("sonar.java.coveragePlugin", "jacoco")
+        property("sonar.java.source", "17")
+        property("sonar.exclusions", sonarExclusions)
+        property("sonar.coverage.exclusions", sonarExclusions)
+
+    }
+}
+tasks.sonar.dependsOn("test", "jacocoTestReport", "jacocoTestCoverageVerification")
