+++ container-start.sh
@@ -1,0 +1,4 @@
+#! /bin/sh
+
+docker-compose -f docker/ibeer-compose.yml up -d && \
+docker logs -f ibeer-service
