--- docker/observability/prometheus/prometheus_data/01H3DY5N3WAV3KC9RWVC5QC316/meta.json
@@ -1,17 +1,0 @@
-{
-	"ulid": "01H3DY5N3WAV3KC9RWVC5QC316",
-	"minTime": 1687238419134,
-	"maxTime": 1687240800000,
-	"stats": {
-		"numSamples": 46741,
-		"numSeries": 703,
-		"numChunks": 713
-	},
-	"compaction": {
-		"level": 1,
-		"sources": [
-			"01H3DY5N3WAV3KC9RWVC5QC316"
-		]
-	},
-	"version": 1
-}
