--- docker/observability/prometheus/prometheus_data/01H3E5VR3CKYRWGK2YYMAXCBQR/meta.json
@@ -1,17 +1,0 @@
-{
-	"ulid": "01H3E5VR3CKYRWGK2YYMAXCBQR",
-	"minTime": 1687313514834,
-	"maxTime": 1687320000000,
-	"stats": {
-		"numSamples": 647606,
-		"numSeries": 1942,
-		"numChunks": 5123
-	},
-	"compaction": {
-		"level": 1,
-		"sources": [
-			"01H3E5VR3CKYRWGK2YYMAXCBQR"
-		]
-	},
-	"version": 1
-}
