+++ src/main/java/com/ipaixao/ibeer/infrastructure/config/BaseMapperConfig.java
@@ -1,0 +1,18 @@
+package com.ipaixao.ibeer.infrastructure.config;
+
+import org.mapstruct.MapperConfig;
+import org.mapstruct.ReportingPolicy;
+
+import static org.mapstruct.InjectionStrategy.CONSTRUCTOR;
+import static org.mapstruct.NullValueCheckStrategy.ON_IMPLICIT_CONVERSION;
+import static org.mapstruct.NullValuePropertyMappingStrategy.IGNORE;
+
+@MapperConfig(
+        componentModel = "spring",
+        injectionStrategy = CONSTRUCTOR,
+        nullValuePropertyMappingStrategy = IGNORE,
+        unmappedSourcePolicy = ReportingPolicy.IGNORE,
+        unmappedTargetPolicy = ReportingPolicy.IGNORE,
+        nullValueCheckStrategy = ON_IMPLICIT_CONVERSION
+)
+public interface BaseMapperConfig {}
