+++ src/main/java/com/ipaixao/ibeer/infrastructure/config/WebSecurityConfig.java
@@ -1,0 +1,39 @@
+package com.ipaixao.ibeer.infrastructure.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
+import org.springframework.security.web.SecurityFilterChain;
+
+import static org.springframework.security.config.http.SessionCreationPolicy.STATELESS;
+
+@Configuration
+@EnableWebSecurity
+public class WebSecurityConfig {
+
+    private static final String[] ALLOWED_URIS;
+
+    static {
+        ALLOWED_URIS = new String[]{
+                "beers/**",
+                "manufacturers/**",
+                "/actuator/**",
+                "/error**",
+                "/v3/api-docs/**",
+                "/api-docs/**",
+                "/v3/api-docs.yaml",
+                "/swagger-ui/**",
+                "/swagger-ui.html"
+        };
+    }
+
+    @Bean
+    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
+        return http.authorizeHttpRequests(authorization -> authorization.requestMatchers(ALLOWED_URIS).permitAll())
+                   .csrf(AbstractHttpConfigurer::disable)
+                   .sessionManagement(session -> session.sessionCreationPolicy(STATELESS))
+                   .build();
+    }
+}
