--- src/main/java/com/ipaixao/ibeer/infrastructure/liquibase/LiquibaseConfig.java
@@ -1,33 +1,0 @@
-package com.ipaixao.ibeer.infrastructure.liquibase;
-
-import liquibase.integration.spring.SpringLiquibase;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.boot.autoconfigure.liquibase.LiquibaseProperties;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import javax.sql.DataSource;
-
-
-@Slf4j
-@Configuration
-@EnableConfigurationProperties(LiquibaseProperties.class)
-public class LiquibaseConfig {
-
-    @Bean
-    public SpringLiquibase liquibase(DataSource dataSource, LiquibaseProperties properties) {
-        log.debug("Configuring Liquibase...");
-
-        final var liquibase = new SpringLiquibase();
-        liquibase.setDataSource(dataSource);
-        liquibase.setChangeLog(properties.getChangeLog());
-        liquibase.setContexts(properties.getContexts());
-        liquibase.setDefaultSchema(properties.getDefaultSchema());
-        liquibase.setDropFirst(properties.isDropFirst());
-        liquibase.setShouldRun(properties.isEnabled());
-
-        log.debug("It's Ok! Liquibase configured.");
-        return liquibase;
-    }
-}
