+++ src/main/java/com/ipaixao/ibeer/interfaces/incomming/beer/mapper/BeerMapper.java
@@ -2,7 +2,7 @@ package com.ipaixao.ibeer.interfaces.incomming.beer.mapper;
 
 import com.ipaixao.ibeer.domain.beer.Beer;
 import com.ipaixao.ibeer.infrastructure.mapstruct.BaseEntityMapper;
-import com.ipaixao.ibeer.infrastructure.mapstruct.BaseMapperConfig;
+import com.ipaixao.ibeer.infrastructure.config.BaseMapperConfig;
 import com.ipaixao.ibeer.infrastructure.mapstruct.BaseResponseMapper;
 import com.ipaixao.ibeer.interfaces.incomming.beer.dto.BeerDTO;
 import com.ipaixao.ibeer.interfaces.incomming.beer.response.BeerResponse;
