+++ src/main/java/com/ipaixao/ibeer/interfaces/incomming/manufacturer/mapper/ManufacturerMapper.java
@@ -1,7 +1,7 @@
 package com.ipaixao.ibeer.interfaces.incomming.manufacturer.mapper;
 
 import com.ipaixao.ibeer.infrastructure.mapstruct.BaseEntityMapper;
-import com.ipaixao.ibeer.infrastructure.mapstruct.BaseMapperConfig;
+import com.ipaixao.ibeer.infrastructure.config.BaseMapperConfig;
 import com.ipaixao.ibeer.infrastructure.mapstruct.BaseResponseMapper;
 import com.ipaixao.ibeer.interfaces.incomming.beer.mapper.BeerMapper;
 import com.ipaixao.ibeer.interfaces.incomming.manufacturer.dto.ManufacturerDTO;
