--- start-app.sh
@@ -1,4 +1,0 @@
-#! /bin/sh
-
-docker-compose -f docker/ibeer-compose.yml up -d && \
-docker logs -f ibeer-app
